fix nesting of queries
